Entorno/Scripts/Tienda/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.4
djangorestframework==3.15.1
filelock==3.13.4
idna==3.7
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
platformdirs==4.2.0
requests==2.31.0
setuptools==69.5.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
